Centralize kernel generation in a private module
Extract box, disk, line, and PSF kernel factory functions into a dedicated `_kernels.py` module. This separates kernel calculation logic from image convolution and simplifies the individual blur implementation files.
