+++ tests/test_linear_motion_blur.py
@@ -7,8 +7,8 @@ from conftest import assert_same_size
 from PIL import Image
 
 import pyblur
+from pyblur._kernels import line_endpoints, line_kernel
 from pyblur._validation import _KERNEL_DIMS
-from pyblur.linear_motion_blur import _line_endpoints, _line_kernel
 
 _LINE_TYPES = ["full", "right", "left"]
 
@@ -17,19 +17,19 @@ class TestLineKernel:
     @pytest.mark.parametrize("dim", _KERNEL_DIMS)
     @pytest.mark.parametrize("linetype", _LINE_TYPES)
     def test_shape(self, dim: int, linetype: str) -> None:
-        k = _line_kernel(dim, 0.0, linetype)  # type: ignore[arg-type]
+        k = line_kernel(dim, 0.0, linetype)  # type: ignore[arg-type]
         assert k.shape == (dim, dim)
 
     @pytest.mark.parametrize("dim", _KERNEL_DIMS)
     @pytest.mark.parametrize("linetype", _LINE_TYPES)
     def test_sums_to_one(self, dim: int, linetype: str) -> None:
-        k = _line_kernel(dim, 45.0, linetype)  # type: ignore[arg-type]
+        k = line_kernel(dim, 45.0, linetype)  # type: ignore[arg-type]
         assert pytest.approx(k.sum(), abs=1e-6) == 1.0
 
     @pytest.mark.parametrize("dim", _KERNEL_DIMS)
     @pytest.mark.parametrize("linetype", _LINE_TYPES)
     def test_non_negative(self, dim: int, linetype: str) -> None:
-        k = _line_kernel(dim, 90.0, linetype)  # type: ignore[arg-type]
+        k = line_kernel(dim, 90.0, linetype)  # type: ignore[arg-type]
         assert (k >= 0).all()
 
 
@@ -147,16 +147,16 @@ class TestLineEndpoints:
     def test_canonical_values(
         self, dim: int, angle: float, expected: tuple[int, int, int, int]
     ) -> None:
-        result = _line_endpoints(dim, angle)
+        result = line_endpoints(dim, angle)
         rev = (expected[2], expected[3], expected[0], expected[1])
         assert result == expected or result == rev, (
-            f"_line_endpoints({dim}, {angle}) = {result}, expected {expected} or {rev}"
+            f"line_endpoints({dim}, {angle}) = {result}, expected {expected} or {rev}"
         )
 
     @pytest.mark.parametrize("dim", [3, 5, 7, 9, 11, 13])
     @pytest.mark.parametrize("angle", [0.0, 22.5, 45.0, 90.0, 135.0, 157.5, 177.3])
     def test_endpoints_within_bounds(self, dim: int, angle: float) -> None:
-        r0, c0, r1, c1 = _line_endpoints(dim, angle)
+        r0, c0, r1, c1 = line_endpoints(dim, angle)
         assert 0 <= r0 < dim and 0 <= c0 < dim
         assert 0 <= r1 < dim and 0 <= c1 < dim
 
@@ -164,16 +164,16 @@ class TestLineEndpoints:
     @pytest.mark.parametrize("angle", [0.0, 45.0, 90.0, 135.0])
     def test_endpoints_symmetric_around_center(self, dim: int, angle: float) -> None:
         c = dim // 2
-        r0, c0, r1, c1 = _line_endpoints(dim, angle)
+        r0, c0, r1, c1 = line_endpoints(dim, angle)
         assert (r0 + r1) // 2 == c or abs((r0 + r1) / 2 - c) < 1.0
         assert (c0 + c1) // 2 == c or abs((c0 + c1) / 2 - c) < 1.0
 
     def test_angle_wrap_360(self) -> None:
-        assert _line_endpoints(9, 0.0) == _line_endpoints(9, 180.0)
+        assert line_endpoints(9, 0.0) == line_endpoints(9, 180.0)
 
     def test_large_dim_shape(self) -> None:
         for dim in [11, 13, 15]:
-            r0, c0, r1, c1 = _line_endpoints(dim, 45.0)
+            r0, c0, r1, c1 = line_endpoints(dim, 45.0)
             assert r0 == dim - 1 and c0 == 0 and r1 == 0 and c1 == dim - 1
 
 
@@ -181,17 +181,17 @@ class TestLineKernelParity:
     """Verify dynamic computation produces the geometrically correct kernel."""
 
     def test_horizontal_kernel(self) -> None:
-        k = _line_kernel(5, 0.0, "full")
+        k = line_kernel(5, 0.0, "full")
         assert np.allclose(k[2], np.full(5, 0.2))
         assert k[:2].sum() == 0 and k[3:].sum() == 0
 
     def test_vertical_kernel(self) -> None:
-        k = _line_kernel(5, 90.0, "full")
+        k = line_kernel(5, 90.0, "full")
         assert np.allclose(k[:, 2], np.full(5, 0.2))
         assert k[:, :2].sum() == 0 and k[:, 3:].sum() == 0
 
     @pytest.mark.parametrize("dim", [11, 13, 15])
     def test_large_dim_sums_to_one(self, dim: int) -> None:
         for angle in [0.0, 45.0, 90.0, 135.0]:
-            k = _line_kernel(dim, angle, "full")
+            k = line_kernel(dim, angle, "full")
             assert pytest.approx(k.sum(), abs=1e-5) == 1.0
